+++ aldrin/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased
+
+## Added
+
+- Add the functions `Discoverer::start_current_only` and `DiscovererBuilder::build_current_only` to
+  have it consider only those object which exist already on the bus at the time of the function
+  call.
+
+
 # 0.1.0 (November 24th, 2023)
 
 - Initial release.